replace package request by service, create a package in order to separate occi service from generic service
