+++ src/main/java/org/ow2/proactive/procci/model/utils/ConvertUtils.java
@@ -21,7 +21,7 @@ public class ConvertUtils {
         try {
             return integerString.map(s -> Integer.parseInt(s));
         } catch (Exception e) {
-            throw new SyntaxException(integerString.get(),"Integer");
+            throw new SyntaxException(integerString.get(), "Integer");
         }
     }
 
@@ -37,7 +37,7 @@ public class ConvertUtils {
         try {
             return floatString.map(s -> Float.parseFloat(s));
         } catch (Exception e) {
-            throw new SyntaxException(floatString.get(),"Float");
+            throw new SyntaxException(floatString.get(), "Float");
         }
     }
 
@@ -49,7 +49,7 @@ public class ConvertUtils {
         try {
             return stringObject.map(obj -> (String) obj);
         } catch (Exception e) {
-            throw new SyntaxException(stringObject.toString(),"String");
+            throw new SyntaxException(stringObject.toString(), "String");
         }
     }
