+++ src/test/java/org/ow2/proactive/procci/model/occi/infrastructure/ComputeTest.java
@@ -12,8 +12,8 @@ import org.ow2.proactive.procci.model.exception.CloudAutomationException;
 import org.ow2.proactive.procci.model.occi.infrastructure.state.ComputeState;
 import org.ow2.proactive.procci.model.occi.metamodel.MixinBuilder;
 import org.ow2.proactive.procci.model.occi.metamodel.rendering.ResourceRendering;
-import org.ow2.proactive.procci.request.CloudAutomationVariablesClient;
-import org.ow2.proactive.procci.request.MixinService;
+import org.ow2.proactive.procci.service.CloudAutomationVariablesClient;
+import org.ow2.proactive.procci.service.occi.MixinService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
