+++ src/test/java/org/ow2/proactive/procci/model/occi/infrastructure/NetworkInterfaceTest.java
@@ -6,8 +6,8 @@ import org.ow2.proactive.procci.model.exception.CloudAutomationException;
 import org.ow2.proactive.procci.model.exception.SyntaxException;
 import org.ow2.proactive.procci.model.occi.infrastructure.constants.InfrastructureKinds;
 import org.ow2.proactive.procci.model.occi.infrastructure.state.NetworkState;
-import org.ow2.proactive.procci.request.CloudAutomationVariablesClient;
-import org.ow2.proactive.procci.request.MixinService;
+import org.ow2.proactive.procci.service.CloudAutomationVariablesClient;
+import org.ow2.proactive.procci.service.occi.MixinService;
 import com.google.common.truth.Truth;
 import org.junit.Before;
 import org.junit.Test;
