+++ src/test/java/org/ow2/proactive/procci/model/occi/metamodel/LinkTest.java
@@ -12,8 +12,8 @@ import org.ow2.proactive.procci.model.occi.infrastructure.Compute;
 import org.ow2.proactive.procci.model.occi.infrastructure.ComputeBuilder;
 import org.ow2.proactive.procci.model.occi.infrastructure.constants.InfrastructureKinds;
 import org.ow2.proactive.procci.model.occi.metamodel.constants.Kinds;
-import org.ow2.proactive.procci.request.CloudAutomationVariablesClient;
-import org.ow2.proactive.procci.request.MixinService;
+import org.ow2.proactive.procci.service.CloudAutomationVariablesClient;
+import org.ow2.proactive.procci.service.occi.MixinService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
