+++ src/test/java/org/ow2/proactive/procci/model/occi/metamodel/ResourceTest.java
@@ -11,8 +11,8 @@ import org.ow2.proactive.procci.model.exception.SyntaxException;
 import org.ow2.proactive.procci.model.occi.infrastructure.Compute;
 import org.ow2.proactive.procci.model.occi.infrastructure.ComputeBuilder;
 import org.ow2.proactive.procci.model.occi.metamodel.constants.Kinds;
-import org.ow2.proactive.procci.request.CloudAutomationVariablesClient;
-import org.ow2.proactive.procci.request.MixinService;
+import org.ow2.proactive.procci.service.CloudAutomationVariablesClient;
+import org.ow2.proactive.procci.service.occi.MixinService;
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mock;
